refactor: 구입금액 유효성 검증하는 부분을 하나의 클래스로 통합
